Move Both uni and dual models to single file
